+++ scm-ui/ui-types/src/Tags.ts
@@ -27,5 +27,6 @@ import { Links } from "./hal";
 export type Tag = {
   name: string;
   revision: string;
+  date?: Date;
   _links: Links;
 };